move logging code to separate file
