+++ cli/internal/terraform/terraform.go
@@ -19,7 +19,6 @@ import (
 	"errors"
 	"fmt"
 	"path/filepath"
-	"strings"
 
 	"github.com/edgelesssys/constellation/v2/internal/cloud/cloudprovider"
 	"github.com/edgelesssys/constellation/v2/internal/constants"
@@ -36,69 +35,9 @@ import (
 )
 
 const (
-	// LogLevelNone represents a log level that does not produce any output.
-	LogLevelNone LogLevel = iota
-	// LogLevelError enables log output at ERROR level.
-	LogLevelError
-	// LogLevelWarn enables log output at WARN level.
-	LogLevelWarn
-	// LogLevelInfo enables log output at INFO level.
-	LogLevelInfo
-	// LogLevelDebug enables log output at DEBUG level.
-	LogLevelDebug
-	// LogLevelTrace enables log output at TRACE level.
-	LogLevelTrace
-	// LogLevelJSON enables log output at TRACE level in JSON format.
-	LogLevelJSON
 	tfVersion         = ">= 1.2.0"
 	terraformVarsFile = "terraform.tfvars"
 )
-
-// LogLevel is a Terraform log level.
-// As per https://developer.hashicorp.com/terraform/internals/debugging
-type LogLevel int
-
-// ParseLogLevel parses a log level string into a Terraform log level.
-func ParseLogLevel(level string) (LogLevel, error) {
-	switch strings.ToUpper(level) {
-	case "NONE":
-		return LogLevelNone, nil
-	case "ERROR":
-		return LogLevelError, nil
-	case "WARN":
-		return LogLevelWarn, nil
-	case "INFO":
-		return LogLevelInfo, nil
-	case "DEBUG":
-		return LogLevelDebug, nil
-	case "TRACE":
-		return LogLevelTrace, nil
-	case "JSON":
-		return LogLevelJSON, nil
-	default:
-		return LogLevelNone, fmt.Errorf("invalid log level %s", level)
-	}
-}
-
-// String returns the string representation of a Terraform log level.
-func (l LogLevel) String() string {
-	switch l {
-	case LogLevelError:
-		return "ERROR"
-	case LogLevelWarn:
-		return "WARN"
-	case LogLevelInfo:
-		return "INFO"
-	case LogLevelDebug:
-		return "DEBUG"
-	case LogLevelTrace:
-		return "TRACE"
-	case LogLevelJSON:
-		return "JSON"
-	default:
-		return ""
-	}
-}
 
 // ErrTerraformWorkspaceExistsWithDifferentVariables is returned when existing Terraform files differ from the version the CLI wants to extract.
 var ErrTerraformWorkspaceExistsWithDifferentVariables = errors.New("creating cluster: a Terraform workspace already exists with different variables")
